add tests for enums and exercise data functions, and remove index.ts
