--- app/src/lib/index.ts
@@ -1,4 +1,0 @@
-// place files you want to import through the `$lib` alias in this folder.
-export * from './enums';
-export * from './types';
-export * from './exerciseData';
